server/requirements.txt
connexion == 1.0.129
python_dateutil == 2.6.0
setuptools >= 21.0.0
flask-cors

